% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bsgenome_from_accession.R
\name{build_bsgenome_from_accession}
\alias{build_bsgenome_from_accession}
\title{Build BSgenome Package from GenBank Accession}
\usage{
build_bsgenome_from_accession(genbank_accession, dir = base::getwd())
}
\arguments{
\item{genbank_accession}{A character string specifying the GenBank accession number (e.g., "GCF_030376745.1").}

\item{dir}{A character string specifying the working directory (default: current working directory).}
}
\value{
A character string with the name of the generated BSgenome package.
}
\description{
Builds a BSgenome package for a given GenBank accession number using a previously downloaded .fna file.
}
\examples{
\dontrun{
pkg_name <- build_bsgenome_from_accession(
  genbank_accession = "GCF_030376745.1",
  dir = getwd()
)
}
}
