% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_deletion_genbank.R
\name{.adjust_one_coord}
\alias{.adjust_one_coord}
\title{Adjust a single coordinate based on insertion site}
\usage{
.adjust_one_coord(coord, ins_start, ins_end, length_diff)
}
\description{
Adjust a single coordinate based on insertion site
}
\keyword{internal}
