% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_deletion_genbank.R
\name{.build_minimal_genbank}
\alias{.build_minimal_genbank}
\title{Build minimal GenBank from name + sequence}
\usage{
.build_minimal_genbank(name, sequence)
}
\description{
Build minimal GenBank from name + sequence
}
\keyword{internal}
