% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_grna_vector_genbank.R
\name{.build_minimal_genbank_grna}
\alias{.build_minimal_genbank_grna}
\title{Build minimal GenBank template for gRNA expression vector}
\usage{
.build_minimal_genbank_grna(name, sequence)
}
\description{
Build minimal GenBank template for gRNA expression vector
}
\keyword{internal}
