% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_deletion_genbank.R
\name{.get_genbank_from_dna}
\alias{.get_genbank_from_dna}
\title{Get GenBank lines from .dna file (via SnapGene CLI)}
\usage{
.get_genbank_from_dna(dna_file, kill_snapgene = FALSE)
}
\description{
Get GenBank lines from .dna file (via SnapGene CLI)
}
\keyword{internal}
