% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_methylation.R
\name{.get_pam_side}
\alias{.get_pam_side}
\title{Get PAM side from nuclease name}
\usage{
.get_pam_side(nuclease)
}
\arguments{
\item{nuclease}{Character. Nuclease name.}
}
\value{
Character. "3prime" or "5prime".
}
\description{
Get PAM side from nuclease name
}
\keyword{internal}
