% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vector.R
\name{.is_snapgene_running}
\alias{.is_snapgene_running}
\title{Check if SnapGene is currently running}
\usage{
.is_snapgene_running()
}
\description{
Check if SnapGene is currently running
}
\keyword{internal}
