% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_methylation.R
\name{.iupac_to_regex}
\alias{.iupac_to_regex}
\title{Convert IUPAC ambiguity sequence to regex pattern}
\usage{
.iupac_to_regex(iupac_seq)
}
\arguments{
\item{iupac_seq}{Character. DNA sequence with IUPAC codes.}
}
\value{
Character. Regex pattern string.
}
\description{
Convert IUPAC ambiguity sequence to regex pattern
}
\keyword{internal}
