% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vector.R
\name{.try_snapgene_cli}
\alias{.try_snapgene_cli}
\title{Try SnapGene CLI conversion to GenBank}
\usage{
.try_snapgene_cli(file_path)
}
\description{
Try SnapGene CLI conversion to GenBank
}
\keyword{internal}
