% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gRNA_for_sequence.R
\name{generate_gRNA_for_sequence}
\alias{generate_gRNA_for_sequence}
\title{Generate gRNA for a Single Sequence}
\usage{
generate_gRNA_for_sequence(
  target_sequence,
  nuclease = "GeoCas9",
  bowtie_index,
  genome,
  position_range = c(10, 90),
  tm_range = c(50, 70),
  strand = "both",
  aligner = "bowtie"
)
}
\arguments{
\item{target_sequence}{A character string specifying the target DNA sequence.}

\item{nuclease}{A character string specifying the nuclease to use ("GeoCas9", "FnCas12a", or "FisCasI_B"; default: "GeoCas9").}

\item{bowtie_index}{A character string specifying the path to the Bowtie index.}

\item{genome}{A BSgenome object specifying the reference genome.}

\item{position_range}{A numeric vector of length 2 specifying the position range (as percentage) for gRNA filtering (default: c(10, 90)).}

\item{tm_range}{A numeric vector of length 2 specifying the Tm range for gRNA filtering (default: c(50, 70)).}

\item{strand}{A character string specifying the strand to filter ("both", "5", or "3"; default: "both").}

\item{aligner}{A character string specifying the aligner to use ("bowtie", "bwa", or "biostrings"; default: "bowtie").}
}
\value{
A data frame containing the gRNA list for the target sequence.
}
\description{
Generates a gRNA list for a single target DNA sequence using the specified nuclease.
}
\examples{
\dontrun{
target_seq <- "ttggcattgacaggtacagaccgcgtcaaacgcggcatggcggaaatgcaaaaaggcggcgtcattatggacgtcgtcaatgcagagcaagcgaagattgctgaggcggcaggggctgtcgcagtcatggcgctcgagcgtgtcccggcagacattcgcgccgctggcggtgtcgcgcgcatggctgatccgacgatcattgaagaagtgatgaacgccgtatcaatcccagtcatggcgaaagtgcgcatcgggcattatgtggaagcgcgtgttttagaggcgctcggcvtcgactatattgacgaaagtgaagtattgacgccggctgatgaagagttccatattgacaaacggcagtttacggtcccatttgtatgcggttgccgcgacttaggagaggccgcccgccgcattgctgaaggggcatcgatgttgcggacaaaaggggagccagggacaggaaacatcgtcgaggccgttcgccatatgcgcaaagtcaacgcgcaaatccgcaaagttgtcagcatgagcgaagacgaacttgtcgccgaggcgaaacagcttggggctccggttgaagtgctgcgtgaaatcaaacggcttggccgcctcccggtcgtcaacttcgccgccggcggtgtcgcgacaccagctgacgccgcgctcatgatgcacttaggcgccgacggtgtctfcgtcggttcgggcatttttaaatcggaaaatccggaaaaatacgctcgtgcgatcgttgaagcgacgactcattatgaagactatgagctgatcgcccatctatcgaaagggctgggcggcgcaatgcgcggcatcgatvtcgcgacactgctgccggagcatcggatgcaagaacgaggctggtaa"
gRNA_list <- generate_gRNA_for_sequence(
  target_sequence = target_seq,
  nuclease = "GeoCas9",
  bowtie_index = "path/to/bowtie_index",
  genome = NULL,
  position_range = c(10, 90),
  tm_range = c(50, 70),
  aligner = "bowtie"
)
print(gRNA_list)
}
}
