% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bowtie_build.R
\name{run_bowtie_build}
\alias{run_bowtie_build}
\title{Build Bowtie Index for a Genome}
\usage{
run_bowtie_build(directory = NULL, accession)
}
\arguments{
\item{directory}{A character string specifying the directory containing the .fna file (default: NULL, uses current working directory).}

\item{accession}{A character string specifying the GenBank accession number (e.g., "GCF_030376745.1").}
}
\value{
None. Creates Bowtie index files in the specified directory.
}
\description{
Builds a Bowtie index for a genome specified by a GenBank accession number.
The index is created from a previously downloaded .fna file.
}
\examples{
\dontrun{
run_bowtie_build(accession = "GCF_030376745.1")
}
}
